Add files via upload
Added some changes that is needed based on the TODO comments in the source codes, mainly the cell.java, board. java and gamemain.java files such as initialising cell rays using rows and cols constraints. In board specificially, I added some code to check for win conditions and check if to see if its a draw or not.

Note that the game is not fully finished as the gamemain java file is still incomplete, needed to add a few more code into it based on the TODO comments.