Add files via upload
Added some more code into the GameMain file, mainly to let the user know if the game is a draw and refreshing the game board if it is, as well as letting the player know who's turn it is in the current round, eg. "It's X's turn" through a message. Gamemain code is done.